Makefile: Run targets in single shell
Currently commands that depend on venv is required to run with same shell
and that is why they are given in single line with ';' as delimiter.

This is also the reason for the hacky way to read the configuration since
reading it is outside the venv context.

GNU Automake has a feature to run everything under single shell as documented
at https://www.gnu.org/software/make/manual/html_node/One-Shell.html
Adding .ONESHELL: somewhere in the Makefile does the trick.
This also slightly improve the performance of targets.